added coding in form.php
After navigating from the welcome.php file, the user will go on form.php. form.php has own .css file which is form.css. and that file connected with 2 files which are form.php and action.php. 
So, after entering the data from the client-side on, the user will be navigated to action.php file which shows the content of successfully added data. and from that file user can go either home or logout or previous page which is form.php.